All endpoints protected - except admin/users/?username=

+++ routes/adminUserRouter.js
@@ -15,13 +15,13 @@ adminUserRouter.route('/')
         res.sendStatus(200);
     })
     .get(cors.cors, (req, res, next) => {
-        User.find(req.query)
-            .then((users) => {
-                res.statusCode = 200;
-                res.setHeader('Content-Type', 'application/json');
-                res.json(users);
-            }, (err) => next(err))
-            .catch((err) => next(err));
+            User.find(req.query)
+                .then((users) => {
+                    res.statusCode = 200;
+                    res.setHeader('Content-Type', 'application/json');
+                    res.json(users);
+                }, (err) => next(err))
+                .catch((err) => next(err));
     })
 ; // end adminUserRouter admin/users/
 
@@ -29,28 +29,42 @@ adminUserRouter.route('/new')
     .options(cors.corsWithOptions, (req, res) => {
         res.sendStatus(200);
     })
-    .get(cors.cors, (req, res, next) => {
-        res.statusCode = 200;
-        res.setHeader('Content-Type', 'text/html');
-        res.send('Howdy');
+    .get(authenticate.verifyUser, cors.cors, (req, res, next) => {
+        if (req.user.isAdmin === true) {
+            res.statusCode = 200;
+            res.setHeader('Content-Type', 'text/html');
+            res.send('Howdy');
+        } else {
+            res.statusCode = 401;
+            res.setHeader('Content-Type', 'application/json');
+            res.json({success: false, message: 'Not authorized'});
+            res.end();
+        }
     })
-    .post(cors.corsWithOptions, (req, res, next) => {
-        User.register(new User({username: req.body.username}),
-            req.body.password, (err, user) => {
-                if(err) {
-                    res.statusCode = 500;
-                    res.setHeader('Content-Type', 'application/json');
-                    res.json({err: err});
-                    console.log(req.body.password);
-                }
-                else {
-                    passport.authenticate('local')(req, res, () => {
-                        res.statusCode = 200;
+    .post(authenticate.verifyUser, cors.corsWithOptions, (req, res, next) => {
+        if (req.user.isAdmin === true) {
+            User.register(new User({username: req.body.username}),
+                req.body.password, (err, user) => {
+                    if (err) {
+                        res.statusCode = 500;
                         res.setHeader('Content-Type', 'application/json');
-                        res.json({success: true, status: 'Registration Successful!'});
-                    });
-                }
-            });
+                        res.json({err: err});
+                        console.log(req.body.password);
+                    }
+                    else {
+                        passport.authenticate('local')(req, res, () => {
+                            res.statusCode = 200;
+                            res.setHeader('Content-Type', 'application/json');
+                            res.json({success: true, status: 'Registration Successful!'});
+                        });
+                    }
+                });
+        } else {
+            res.statusCode = 401;
+            res.setHeader('Content-Type', 'application/json');
+            res.json({success: false, message: 'Not authorized'});
+            res.end();
+        }
     })
 ; // end adminUserRouter admin/users/new
 
@@ -58,28 +72,42 @@ adminUserRouter.route('/:userId')
     .options(cors.corsWithOptions, (req, res) => {
         res.sendStatus(200);
     })
-    .get(cors.cors, (req, res, next) => {
-        User.findById(req.params.userId)
-            .then((user) => {
-                res.statusCode = 200;
-                res.setHeader('Content-Type', 'application/json');
-                res.json(user);
-            }, (err) => next(err))
-            .catch((err) => next(err));
+    .get(authenticate.verifyUser, cors.cors, (req, res, next) => {
+        if (req.user.isAdmin === true) {
+            User.findById(req.params.userId)
+                .then((user) => {
+                    res.statusCode = 200;
+                    res.setHeader('Content-Type', 'application/json');
+                    res.json(user);
+                }, (err) => next(err))
+                .catch((err) => next(err));
+        } else {
+            res.statusCode = 401;
+            res.setHeader('Content-Type', 'application/json');
+            res.json({success: false, message: 'Not authorized'});
+            res.end();
+        }
     })
     .post(cors.corsWithOptions, (req, res, next) => {
         res.statusCode = 403;
         res.end('POST operation not supported on /users/' + req.params.userId);
     })
 
-    .delete(cors.corsWithOptions, (req, res, next) => {
-        User.findByIdAndRemove(req.params.userId)
-            .then((resp) => {
-                res.statusCode = 200;
-                res.setHeader('Content-Type', 'application/json');
-                res.json(resp);
-            }, (err) => next(err))
-            .catch((err) => next(err));
+    .delete(authenticate.verifyUser, cors.corsWithOptions, (req, res, next) => {
+        if (req.user.isAdmin === true) {
+            User.findByIdAndRemove(req.params.userId)
+                .then((resp) => {
+                    res.statusCode = 200;
+                    res.setHeader('Content-Type', 'application/json');
+                    res.json(resp);
+                }, (err) => next(err))
+                .catch((err) => next(err));
+        } else {
+            res.statusCode = 401;
+            res.setHeader('Content-Type', 'application/json');
+            res.json({success: false, message: 'Not authorized'});
+            res.end();
+        }
     })
 ; // end adminUserRouter admin/users/:userId
 
@@ -87,41 +115,61 @@ adminUserRouter.route('/edit/:userId')
     .options(cors.corsWithOptions, (req, res) => {
         res.sendStatus(200);
     })
-    .get(cors.cors, (req, res, next) => {
-        User.findById(req.params.userId)
-            .then((user) => {
-                res.statusCode = 200;
-                res.setHeader('Content-Type', 'application/json');
-                res.json(user);
-            }, (err) => next(err))
-            .catch((err) => next(err));
+    .get(authenticate.verifyUser, cors.cors, (req, res, next) => {
+        if (req.user.isAdmin === true) {
+            User.findById(req.params.userId)
+                .then((user) => {
+                    res.statusCode = 200;
+                    res.setHeader('Content-Type', 'application/json');
+                    res.json(user);
+                }, (err) => next(err))
+                .catch((err) => next(err));
+        } else {
+            res.statusCode = 401;
+            res.setHeader('Content-Type', 'application/json');
+            res.json({success: false, message: 'Not authorized'});
+            res.end();
+        }
     })
     .post(cors.corsWithOptions, (req, res, next) => {
         res.statusCode = 403;
         res.end('POST operation not supported on /users/edit/' + req.params.userId);
     })
-    .put(cors.corsWithOptions, (req, res, next) => {
-        User.findByIdAndUpdate(req.params.userId, {
-            $set: req.body
-        }, {new: true})
-            .then((user) => {
-                res.statusCode = 200;
-                res.setHeader('Content-Type', 'application/json');
-                res.json(user);
-            }, (err) => next(err))
-            .catch((err) => next(err));
+    .put(authenticate.verifyUser, cors.corsWithOptions, (req, res, next) => {
+        if (req.user.isAdmin === true) {
+            User.findByIdAndUpdate(req.params.userId, {
+                $set: req.body
+            }, {new: true})
+                .then((user) => {
+                    res.statusCode = 200;
+                    res.setHeader('Content-Type', 'application/json');
+                    res.json(user);
+                }, (err) => next(err))
+                .catch((err) => next(err));
+        } else {
+            res.statusCode = 401;
+            res.setHeader('Content-Type', 'application/json');
+            res.json({success: false, message: 'Not authorized'});
+            res.end();
+        }
     })
-    .delete(cors.corsWithOptions, (req, res, next) => {
-        User.findByIdAndRemove(req.params.userId)
-            .then((resp) => {
-                res.statusCode = 200;
-                res.setHeader('Content-Type', 'application/json');
-                res.json(resp);
-            }, (err) => next(err))
-            .catch((err) => next(err));
+    .delete(authenticate.verifyUser, cors.corsWithOptions, (req, res, next) => {
+        if (req.user.isAdmin === true) {
+            User.findByIdAndRemove(req.params.userId)
+                .then((resp) => {
+                    res.statusCode = 200;
+                    res.setHeader('Content-Type', 'application/json');
+                    res.json(resp);
+                }, (err) => next(err))
+                .catch((err) => next(err));
+        } else {
+            res.statusCode = 401;
+            res.setHeader('Content-Type', 'application/json');
+            res.json({success: false, message: 'Not authorized'});
+            res.end();
+        }
     });
 ; // end adminUserRouter admin/users/edit/:userId
-
 
 
 module.exports = adminUserRouter;
